requirements.txt
Flask~=2.2.2
mysql-connector-python~=8.0.30
pandas~=1.5.0
jellyfish~=0.9.0
pysftp~=0.2.9
python-dotenv